Move common metadata operations and plumbing to DSpaceObject.
